ajout de time_exe pour le debug et des printf de malloc

+++ includes/unknow_project_mesh.h
@@ -6,7 +6,7 @@
 
 typedef struct		s_door
 {
-	float			tick;
+	int				tick;
 	int				state;
 	int				move;
 }					t_door;
